reflactor session and context flow
