+++ include/aquarius.hpp
@@ -2,10 +2,10 @@
 #define BOOST_BIND_GLOBAL_PLACEHOLDERS
 #include <aquarius/client.hpp>
 #include <aquarius/connect.hpp>
-#include <aquarius/server.hpp>
 #include <aquarius/context.hpp>
 #include <aquarius/request.hpp>
 #include <aquarius/response.hpp>
+#include <aquarius/server.hpp>
 #include <aquarius/service.hpp>
 
 namespace aquarius
