+++ include/aquarius/context/context.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <aquarius/context/impl/visitor.hpp>
 #include <aquarius/core/logger.hpp>
-#include <aquarius/session/session.hpp>
+#include <aquarius/session/basic_session.hpp>
 #include <chrono>
 
 using namespace std::chrono_literals;
@@ -34,8 +34,7 @@ namespace aquarius
 		basic_context& operator=(const basic_context&) = delete;
 
 	public:
-		virtual error_code visit(basic_message*, std::shared_ptr<basic_session> session_ptr,
-								 error_code& ec) override
+		virtual error_code visit(basic_message*, std::shared_ptr<basic_session> session_ptr, error_code& ec) override
 		{
 			XLOG_WARNING() << name_ << " maybe visit an unknown message!";
 
@@ -49,11 +48,6 @@ namespace aquarius
 
 	public:
 		virtual void on_accept()
-		{
-			return;
-		}
-
-		virtual void on_close()
 		{
 			return;
 		}
