+++ test/unit_test.vcxproj
@@ -182,7 +182,7 @@
     <ClInclude Include="..\include\aquarius\router\message.hpp" />
     <ClInclude Include="..\include\aquarius\router\service.hpp" />
     <ClInclude Include="..\include\aquarius\router\session.hpp" />
-    <ClInclude Include="..\include\aquarius\session\session.hpp" />
+    <ClInclude Include="..\include\aquarius\session\basic_session.hpp" />
     <ClInclude Include="..\include\aquarius\core\core.hpp" />
     <ClInclude Include="..\include\aquarius\core\elastic.hpp" />
     <ClInclude Include="..\include\aquarius\invoke\context.hpp" />
@@ -196,7 +196,7 @@
     <ClInclude Include="..\include\aquarius\response.hpp" />
     <ClInclude Include="..\include\aquarius\server.hpp" />
     <ClInclude Include="..\include\aquarius\service.hpp" />
-    <ClInclude Include="..\include\aquarius\session.hpp" />
+    <ClInclude Include="..\include\aquarius\session\session.hpp" />
     <ClInclude Include="..\include\aquarius\core\asio.hpp" />
     <ClInclude Include="..\include\aquarius\core\deadline_timer.hpp" />
     <ClInclude Include="..\include\aquarius\core\empty_value.hpp" />
